Use the custom hook for slidesPerView
